requirements.txt
flask
pygame
opencv-python
numpy
pillow
psutil
requests
zeroconf
gpiozero
spidev
rpi-ws281x
qrcode[pil]
pyzbar
pydub
